refactor: FilterConnection component, clean up render
